@@ -5,81 +5,75 @@
   "dependencies": {
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
-    "@mdi/js": "^5.9.55",
-    "@mdi/react": "^1.4.0",
-    "axios": "^0.21.1",
-    "express": "^4.17.1",
-    "express-static-gzip": "^2.0.8",
+    "@mdi/js": "^6.5.95",
+    "@mdi/react": "^1.5.0",
+    "axios": "^0.24.0",
+    "clean-webpack-plugin": "^4.0.0",
     "history": "^4.10.1",
-    "i18next": "^19.4.3",
-    "install": "^0.13.0",
-    "keycloak-js": "^11.0.3",
+    "keycloak-js": "^15.0.2",
     "leaflet": "^1.7.1",
     "moment": "^2.29.1",
-    "moment-timezone": "^0.5.27",
+    "moment-timezone": "^0.5.34",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-dropzone": "^11.2.4",
-    "react-leaflet": "^3.1.0",
-    "react-redux": "^7.2.2",
-    "react-router": "^5.1.8",
-    "react-router-dom": "^5.1.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-dropzone": "^11.4.2",
+    "react-leaflet": "^3.2.2",
+    "react-redux": "^7.2.6",
+    "react-router": "^6.0.2",
+    "react-router-dom": "^6.0.2",
     "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-saga": "^1.1.3",
-    "utm": "^1.1.1"
+    "utm": "^1.1.1",
+    "webpack-merge": "^5.8.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.2",
-    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.2",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-react": "^7.13.13",
-    "@babel/preset-typescript": "^7.13.0",
-    "@types/node": "^14.14.11",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-leaflet": "^2.5.2",
-    "@types/react-redux": "^7.1.16",
-    "@types/react-router": "^5.1.8",
-    "@types/react-router-dom": "^5.1.6",
-    "@typescript-eslint/eslint-plugin": "^2.11.0",
-    "@typescript-eslint/parser": "^2.11.0",
-    "babel-eslint": "^10.0.3",
-    "babel-jest": "^24.9.0",
-    "babel-loader": "^8.2.2",
-    "compression-webpack-plugin": "^5.0.2",
-    "core-js": "^3.13.0",
-    "css-loader": "^3.3.0",
-    "dotenv": "^8.1.0",
-    "eslint": "^6.7.2",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-plugin-import": "^2.19.1",
-    "eslint-plugin-jest": "^22.21.0",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-node": "^10.0.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.17.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "eslint-plugin-standard": "^4.0.1",
-    "file-loader": "^5.0.2",
-    "html-webpack-plugin": "^4.5.2",
-    "jest": "^24.9.0",
+    "@babel/core": "^7.16.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.16.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.13",
+    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
+    "@babel/preset-env": "^7.16.4",
+    "@babel/preset-react": "^7.16.0",
+    "@babel/preset-typescript": "^7.16.0",
+    "@types/node": "^16.11.12",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-leaflet": "^2.8.2",
+    "@types/react-redux": "^7.1.20",
+    "@types/react-router": "^5.1.17",
+    "@types/react-router-dom": "^5.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.6.0",
+    "@typescript-eslint/parser": "^5.6.0",
+    "@webpack-cli/serve": "^1.6.0",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^27.4.2",
+    "babel-loader": "^8.2.3",
+    "compression-webpack-plugin": "^9.0.1",
+    "core-js": "^3.19.3",
+    "css-loader": "^6.5.1",
+    "eslint": "^8.4.1",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-standard": "^5.0.0",
+    "html-webpack-plugin": "^5.5.0",
+    "jest": "^27.4.3",
     "mini-css-extract-plugin": "^0.12.0",
-    "node-sass": "^4.13.1",
-    "regenerator-runtime": "^0.13.7",
-    "sass-loader": "^8.0.2",
-    "ts-node": "^9.1.1",
-    "typescript": "^3.9.9",
-    "url-loader": "^3.0.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12"
+    "node-sass": "^7.0.0",
+    "regenerator-runtime": "^0.13.9",
+    "resolve-url-loader": "^4.0.0",
+    "sass-loader": "^12.4.0",
+    "style-loader": "^3.3.1",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.2",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.6.0"
   },
   "scripts": {
-    "dist": "webpack --production --config webpack.dist.js",
-    "build": "webpack --production --config webpack.dist.js",
+    "dist": "webpack --mode production --config webpack.dist.js",
+    "build": "webpack --mode production --config webpack.dist.js",
+    "dev": "webpack serve --config webpack.dev.js ",
     "check-types": "tsc",
     "test": "jest --coverage",
     "lint": "eslint '*/**/*.{js,ts,tsx}'"
