@@ -1,12 +1,11 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import ListComponent from "../../components/ListComponent";
-import {useDispatch, useSelector} from "react-redux";
-import {PermitActions, ReportActions} from "../../../state/actions";
+import {useSelector} from "react-redux";
+import {PermitActions} from "../../../state/actions";
 import Loading from "../../components/Loading";
 import ButtonBar from "../../components/ButtonBar";
 import {Button} from "@material-ui/core";
-import {useHistory} from "react-router-dom";
-import moment from "moment";
+import {useNavigate} from "react-router-dom";
 import {useList} from "../../../state/utilities/use_list";
 import FriendlyTime from "../../components/FriendlyTime";
 
@@ -14,6 +13,7 @@ const PermitList = () => {
 
   const items = useSelector(state => state.Permits.items);
   const loading = useSelector(state => state.Permits.loading);
+  const navigate = useNavigate();
 
   useList(PermitActions, 'area_admin');
 
@@ -25,8 +25,8 @@ const PermitList = () => {
     [
       <td key='org'>{it.organizationname}</td>,
       <td key='ref'>{it.reference}</td>,
-      <td key='sd'><FriendlyTime value={it.startdate}/></td>,
-      <td key='ed'><FriendlyTime value={it.enddate}/></td>
+      <td key='sd'><FriendlyTime value={it.startdate} /></td>,
+      <td key='ed'><FriendlyTime value={it.enddate} /></td>
     ]
   )
   return (
@@ -37,7 +37,7 @@ const PermitList = () => {
         <Button
           variant="contained"
           color="primary"
-          onClick={() => history.push('/area_admin/permits/add')}
+          onClick={() => navigate('/area_admin/permits/add')}
         >Create New Assignment</Button>
       </ButtonBar>
 
