@@ -1,12 +1,11 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import ListComponent from "../../components/ListComponent";
-import {useDispatch, useSelector} from "react-redux";
-import {ReportActions, TenureActions} from "../../../state/actions";
+import {useSelector} from "react-redux";
+import {TenureActions} from "../../../state/actions";
 import Loading from "../../components/Loading";
 import ButtonBar from "../../components/ButtonBar";
 import {Button} from "@material-ui/core";
-import {useHistory} from "react-router-dom";
-import moment from "moment";
+import {useNavigate} from "react-router-dom";
 import {useList} from "../../../state/utilities/use_list";
 import FriendlyTime from "../../components/FriendlyTime";
 
@@ -14,7 +13,7 @@ const TenureList = () => {
 
   const items = useSelector(state => state.Tenures.items);
   const loading = useSelector(state => state.Tenures.loading);
-  const history = useHistory();
+  const navigate = useNavigate();
 
   useList(TenureActions, 'license_auth_officer');
 
@@ -27,8 +26,8 @@ const TenureList = () => {
       <td key='n'>{it.organizationname}</td>,
       <td key='ref'>{it.reference}</td>,
       <td key='st'>{it.subtenures}</td>,
-      <td key='sd'><FriendlyTime value={it.startdate}/></td>,
-      <td key='ed'><FriendlyTime value={it.enddate}/></td>
+      <td key='sd'><FriendlyTime value={it.startdate} /></td>,
+      <td key='ed'><FriendlyTime value={it.enddate} /></td>
     ]
   )
 
@@ -39,7 +38,7 @@ const TenureList = () => {
         <Button
           variant="contained"
           color="primary"
-          onClick={() => history.push('/license_auth_officer/tenures/add')}
+          onClick={() => navigate('/license_auth_officer/tenures/add')}
         >Create New Assignment</Button>
       </ButtonBar>
 
